Add Unit test

+++ OnlineShop/Data/DeleteOrderLineArgs.cs
@@ -2,7 +2,7 @@ namespace OnlineShop.Data
 {       
     public class DeleteOrderLineArgs
     {
-        public int OrderId {get; set; }
-        public int OrderLineId { get; set; }
+        public long OrderId {get; set; }
+        public long OrderLineId { get; set; }
     }
 }
